backend/generic3/requirements.txt
Django==4.2
djangorestframework==3.14.0
djangorestframework-simplejwt
django-cors-headers==3.11.0
django-otp==1.3.0
django-otp[email,totp]
python-dotenv
dj-database-url
psycopg2-binary
gunicorn
whitenoise
